+++ packages/create-app/src/lib/utils/getPkgManager.ts
@@ -9,10 +9,17 @@ export function getPkgManager() {
   if (fromUserAgent) {
     return fromUserAgent.name;
   }
+  return getPkgManagerFromLockFile();
+}
+
+export function getPkgManagerFromLockFile() {
   if (fs.existsSync(path.join(process.cwd(), 'pnpm-lock.yaml'))) {
     return 'pnpm';
   }
-  if (fs.existsSync(path.join(process.cwd(), 'yarn.lock'))) {
+  if (
+    fs.existsSync(path.join(process.cwd(), 'yarn.lock')) ||
+    fs.existsSync(path.join(process.cwd(), '.yarn'))
+  ) {
     return 'yarn';
   }
   if (