+++ packages/provider-github/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @rnef/provider-github
 
+## 0.9.2
+
+### Patch Changes
+
+- b80451d: chore: update no git remote warning
+- Updated dependencies [b80451d]
+- Updated dependencies [1f0999d]
+- Updated dependencies [8180d73]
+  - @rock-js/tools@0.9.2
+
 ## 0.9.1
 
 ### Patch Changes